backend/requirements.txt
FastAPI==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
PyPDF2==3.0.1
python-docx==1.1.2
transformers==4.46.0
torch==2.5.1
numpy==1.26.4
scikit-learn==1.5.2
aiofiles==24.1.0
pydantic==2.9.2
python-dotenv==1.0.1
